requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
django-autocomplete-light==3.4.1
django-bootstrap4==1.0.1
django-cors-headers==3.1.0
django-crontab==0.7.1
django-filter==2.2.0
django-formtools==2.1
django-froala-editor==3.0.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.2
idna==2.8
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
numpy==1.18.1
pandas==0.25.3
Pillow==6.1.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
six==1.12.0
slugify==0.0.1
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wrapt==1.11.2
